fix(cluster/info): apply review feedback
- schema.go: fix three typos ported from SDK (Elasticsearh, Minium x2)
- data_source.go: replace unreachable int64 arm in BuildDate type-switch
  with float64 (JSON numbers decode to float64, not int64)
- acc_test.go: strengthen coverage per schema-coverage review:
  - add tagline exact-value assertion to _topLevelAttributes
  - add build_snapshot == "false" concrete assertion to _versionFieldFormats
  - add headers absent assertion to _withExplicitConnection
- docs: regenerate docs/data-sources/elasticsearch_info.md (typo fixes,
  version sub-field descriptions, Attributes List type)

+++ internal/elasticsearch/cluster/info/data_source.go
@@ -79,8 +79,9 @@ func readDataSource(ctx context.Context, esClient *clients.ElasticsearchScopedCl
 	switch v := res.Version.BuildDate.(type) {
 	case string:
 		buildDate = v
-	case int64:
-		buildDate = fmt.Sprintf("%d", v)
+	case float64:
+		// JSON numbers decode to float64; format as integer milliseconds string.
+		buildDate = fmt.Sprintf("%d", int64(v))
 	default:
 		buildDate = fmt.Sprintf("%v", v)
 	}
